+++ common/src/main/java/io/github/flemmli97/improvedmobs/mixinhelper/CachedRawPathTypes.java
@@ -1,0 +1,9 @@
+package io.github.flemmli97.improvedmobs.mixinhelper;
+
+import it.unimi.dsi.fastutil.longs.Long2ObjectMap;
+import net.minecraft.world.level.pathfinder.BlockPathTypes;
+
+public interface CachedRawPathTypes {
+
+    Long2ObjectMap<BlockPathTypes> getPathTypeByPosCacheRaw();
+}